Created hook to add film to user list
